Recover four stubs via the Mac oracle, and fix TFuzzyVar's field type
The Mac CodeWarrior symbol list turns out to be a precise oracle for these
stubs: cross-referencing each stub's owning class against the Mac names that
are not yet declared anywhere in our source narrows most classes to one or two
candidates, and the bodies then confirm which is which.

0x00551820 TAdmiral::Victory
0x00551990 TAdmiral::IsSeniorTo
  TAdmiral had exactly two undeclared Mac names, and the two stubs in its range
  match them unambiguously: one adds to experiencePoints and clamps the total to
  499 (keeping the derived tier, experiencePoints / 100, within four bands), the
  other compares two admirals' experience. IsSeniorTo tests its own receiver for
  null, so a null admiral ranks below everything -- reproduced as written.

0x004b6a00 TPopulationMgr::AddUntrained
  Mirrors the adjacent AddExpert (0x4b6a30) but writes TLaborPool::lowSkillCount04
  instead of highSkillCount08, and adds no strength -- untrained workers
  contribute none. AddUntrained was the class's one undeclared Mac name.

0x004ff520 TFuzzyVar::IFuzzyVar
  Stores the four fuzzy breakpoints.

TFuzzyVar's values[4] were modelled as int, but every reader touches them as
`float ptr [ecx+N]` -- Membership (0x4ff550/0x4ff5f0) does FCOMP float ptr
[ecx+4]. Correcting the type to float fixes real integer arithmetic in
TFuzzySet::SelectWeightedMemberIndex, whose interpolation
(input - values[0]) / (values[1] - values[0]) was being computed in ints:

  0x4ff840 TFuzzySet::SelectWeightedMemberIndex  35.92% -> 81.94%  (+46pp)

That change also exposed the constants feeding the set. They were written as raw
IEEE-754 bit patterns (0xccbebc20, 0xc7c35000, ...) because the fields were ints;
each decodes to a round number, so they are now written as the float literals
they always were:

  (-1e8, -100000, -20000, -10000) (-15000, -5000, -5000, 1000)
  (0, 5000, 10000, 15000)         (10000, 20000, 1e6, 1e9)

The last of those was `kControlTagNnkParen`, a supposed four-character control
tag. It is not a tag at all: 0x4e6e6b28 is the float 1000000000.0, whose bit
pattern happens to spell "Nnk(". It had exactly one use -- this call -- so the
constant is removed and the fourcc inventory drops from 552 to 551 tags.

Verified: just precommit exit 0. average similarity 77.96% -> 78.02% (+0.06pp),
exact/original 51.38% -> 51.40% (+0.02pp), 16 improved, 11 lower (untouched code,
the usual linking wobble from adding functions across several TUs).

+++ src/game/gfx/TFuzzySet_AllocateAndAppendRecord.cpp
@@ -13,7 +13,7 @@
 // TFuzzyVar; its receiver is a 0x30-byte TFuzzySet (count@0x4 + m_members[10]@0x8),
 // not the 0x14-byte TFuzzyVar leaf it allocates.
 // FUNCTION: IMPERIALISM 0x004ff7d0
-void TFuzzySet::AllocateAndAppendRecord(int value0, int value1, int value2, int value3) {
+void TFuzzySet::AllocateAndAppendRecord(float value0, float value1, float value2, float value3) {
   TFuzzyVar* record = new TFuzzyVar();
   record->values[0] = value0;
   record->values[1] = value1;
